Extract column menu logic to new file.
